Changed: bye content-loader, hello phenomic-loader & phenomic-loader-plugin/presets 🚀
