+++ .flowconfig
@@ -13,3 +13,5 @@
 [options]
 esproposal.class_instance_fields=enable
 esproposal.class_static_fields=enable
+
+suppress_comment= \\(.\\|\n\\)*\\$FlowFixMe
