+++ docs/content/docs/usage/feeds.md
@@ -3,18 +3,23 @@ title: How to use generate feeds in Phenomic
 incomplete: true
 ---
 
-The `phenomic/lib/content-loader` can handle this very easily.
+The ``phenomicLoader`` allows you to generate RSS feeds very easily.
+
 The default boilerplate already handles a default RSS feed that will grab the
 last 20 posts.
 
 Here is a commented part of the webpack configuration that will help:
 
 ```js
+import { phenomicLoader } from "phenomic"
+
+// ...
+
   module: {
    loaders: [
-     { // phenomic requirement
+     {
        test: /\.md$/,
-       loader: contentLoader,
+       loader: phenomicLoader,
        query: {
          context: path.join(config.cwd, config.source),
 
@@ -37,5 +42,5 @@ Here is a commented part of the webpack configuration that will help:
          },
        },
      },
-  ...
+  // ...
 ```
