+++ docs/webpack.config.babel.js
@@ -2,7 +2,7 @@ import path from "path"
 
 import webpack from "webpack"
 import ExtractTextPlugin from "extract-text-webpack-plugin"
-import { contentLoader } from "phenomic"
+import { phenomicLoader } from "phenomic"
 
 import pkg from "./package.json"
 
@@ -21,8 +21,8 @@ export const makeConfig = (config = {}) => {
         {
           // phenomic requirement
           test: /\.md$/,
-          loader: contentLoader,
-          // config is in phenomic.contentLoader section below
+          loader: phenomicLoader,
+          // config is in phenomic.phenomicLoader section below
           // so you can use functions (and not just JSON) due to a restriction
           // of webpack that serialize/deserialize loader `query` option.
         },
@@ -84,21 +84,20 @@ export const makeConfig = (config = {}) => {
     },
 
     phenomic: {
-      contentLoader: {
-        context: path.join(__dirname, config.source),
-        // renderer: (text) => html
-        feedsOptions: {
-          title: pkg.name,
-          site_url: pkg.homepage,
-        },
-        feeds: {
-          "feed.xml": {
-            collectionOptions: {
-              filter: { layout: "Post" },
-              sort: "date",
-              reverse: true,
-              limit: 20,
-            },
+      context: path.join(__dirname, config.source),
+      // plugins: [ ...phenomicLoaderPresets.markdown ]
+      // see https://phenomic.io/docs/usage/plugins/
+      feedsOptions: {
+        title: pkg.name,
+        site_url: pkg.homepage,
+      },
+      feeds: {
+        "feed.xml": {
+          collectionOptions: {
+            filter: { layout: "Post" },
+            sort: "date",
+            reverse: true,
+            limit: 20,
           },
         },
       },
