+++ package.json
@@ -72,7 +72,6 @@
     "simple-json-fetch": "^1.0.1",
     "source-map-support": "^0.4.0",
     "strip-markdown": "^0.3.1",
-    "underscore.string": "^3.2.3",
     "url-join": "^1.1.0",
     "valid-url": "^1.0.9",
     "webpack": "^1.11.0",
