+++ src/builder/server.js
@@ -10,10 +10,10 @@ import opn from "opn"
 import debug from "debug"
 import portFinder from "portfinder"
 
-import minifyCollection from "../content-loader/minify"
+import minifyCollection from "../phenomic-loader/minify"
 import serialize from "../_utils/serialize"
 
-import collection from "../content-loader/cache.js"
+import collection from "../phenomic-loader/cache.js"
 import pathToUri from "../_utils/path-to-uri"
 
 const log = debug("phenomic:builder:server")
