--- src/content-loader-plugins/add-head-property-from-config/index.js
@@ -1,13 +1,0 @@
-// @flow
-
-export default (
-  {
-    result,
-    options,
-  }: PhenomicContentLoaderPluginInput
-): PhenomicCollectionItem => {
-  return {
-    ...result,
-    ...options.defaultHead && { head: options.defaultHead },
-  }
-}
