--- src/content-loader-plugins/index.js
@@ -1,9 +1,0 @@
-/* eslint-disable max-len */
-export addBodyPropertyFromContent from "./add-body-property-from-content"
-export addHeadPropertyFromConfig from "./add-head-property-from-config"
-export addHeadPropertyFromContent from "./add-head-property-from-content"
-export addHeadDescriptionPropertyFromMdContent from "./add-head.description-property-from-md-content"
-export addRawPropertyFromContent from "./add-raw-property-from-content"
-export addRawBodyPropertyFromContent from "./add-rawBody-property-from-content"
-export transformMdBodyPropertyToHtml from "./transform-md-body-property-to-html"
-/* eslint-enable max-len */
