+++ src/phenomic-loader-plugin-init-body-property-from-content/__tests__/index.js
