+++ src/phenomic-loader-plugin-init-head-property-from-config/__tests__/index.js
@@ -1,0 +1,45 @@
+import test from "ava"
+
+import plugin from ".."
+
+test("should add head property from config", (t) => {
+  t.deepEqual(
+    plugin({
+      result: {},
+      options: {
+        defaultHead: {
+          "te": "st",
+        },
+      },
+    }),
+    {
+      head: {
+        "te": "st",
+      },
+    }
+  )
+})
+
+test("should not override existing head", (t) => {
+  t.deepEqual(
+    plugin({
+      result: {
+        head: {
+          te: "sla",
+        },
+      },
+      options: {
+        defaultHead: {
+          "tes": "t",
+          "te": "st",
+        },
+      },
+    }),
+    {
+      head: {
+        "te": "sla",
+        "tes": "t",
+      },
+    }
+  )
+})
