+++ src/phenomic-loader-plugin-init-head-property-from-config/index.js
@@ -1,0 +1,18 @@
+// @flow
+
+export default (
+  {
+    result,
+    options,
+  }: PhenomicLoaderPluginInput
+): PhenomicCollectionItem => {
+  return {
+    ...result,
+    ...options && options.defaultHead && {
+      head: {
+        ...options.defaultHead,
+        ...result.head,
+      },
+    },
+  }
+}
