+++ src/phenomic-loader-plugin-init-head-property-from-content/__tests__/index.js
