+++ src/phenomic-loader-preset-default/__tests__/index.js
@@ -1,0 +1,21 @@
+import test from "ava"
+
+import preset from ".."
+
+test("should do the basic requirements", (t) => {
+  t.deepEqual(
+    preset.reduce((result, plugin) => plugin({
+      frontMatter: {
+        orig: "---\ntitle: Front Matter\n---\nThis is content.",
+        data: { title: "Front Matter" },
+        content: "\nThis is content.",
+      },
+      result,
+      options: {},
+    }), {}),
+    {
+      head: { title: "Front Matter" },
+      body: "\nThis is content.",
+    }
+  )
+})
