+++ src/phenomic-loader/__tests__/fixtures/script.js
